refactor: rename several plugin components to follow same structure
